Create some elements classes (switch, link, port, address) and put some basic code (getter-setter)

+++ src/org/openflow/util/U8.java
@@ -19,7 +19,7 @@ package org.openflow.util;
 
 public class U8 {
     public static short f(byte i) {
-        return (short) ((short)i & 0xff);
+        return (short) (i & 0xff);
     }
 
     public static byte t(short l) {
